requirements.txt
django-cors-headers==2.5.2
django-filter==1.1.0
djangorestframework==3.9.2
